Rename package to teraranger_array
