+++ include/teraranger_array/helper_lib.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-namespace teraranger_hub
+namespace teraranger_array
 {
 static const uint8_t crc_table[] = {0x00, 0x07, 0x0e, 0x09, 0x1c, 0x1b, 0x12, 0x15, 0x38, 0x3f, 0x36, 0x31, 0x24, 0x23,
                                     0x2a, 0x2d, 0x70, 0x77, 0x7e, 0x79, 0x6c, 0x6b, 0x62, 0x65, 0x48, 0x4f, 0x46, 0x41,
@@ -31,6 +31,6 @@ class HelperLib
         static uint8_t crc8(uint8_t *p, uint8_t len);
         static float two_chars_to_float(uint8_t c1, uint8_t c2);
 };
-} // namespace teraranger_hub
+} // namespace teraranger_array
 
 #endif // TERARANGER_HELPER_LIB_H
