+++ package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package>
-  <name>teraranger_hub</name>
+  <name>teraranger_array</name>
   <version>1.0.0</version>
-  <description>The teraranger_hub package</description>
+  <description>The teraranger_array package</description>
 
   <maintainer email="pierre-louis.kabaradjian@terabee.com">Pierre-Louis Kabaradjian</maintainer>
   <maintainer email="krzysztof.zurad@terabee.com">Krzysztof Zurad</maintainer>
